+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Improved sixth (6th) game.
+- Added warning error if translation key doesn't exist.
 
 ### Fixed